Update configuration vue render
