+++ 03 Render/src/hello.ts
@@ -1,0 +1,14 @@
+import Vue from 'vue';
+
+export const HelloComponent = Vue.extend({
+  template: `
+    <input
+      :value="message"
+      @input="onChange($event.target.value)"
+    />
+  `,
+  props: {
+    message: String,
+    onChange: Function,
+  },
+});
